requirements.txt
#先把pip版本降到24.1以下
pip==24.0

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# 数据库和图数据库
neo4j==5.15.0
py2neo==2021.2.4

# 机器学习和NLP
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.0.3

# HTTP请求
requests==2.31.0
httpx==0.25.2

# 配置管理
python-dotenv==1.0.0
PyYAML==6.0.1

# 火山引擎SDK
volcengine-python-sdk[ark]>=1.0.0

# 日志和监控
loguru==0.7.2

# 缓存
redis==5.0.1

# 数据验证
pydantic==1.10.12

# 时间处理
python-dateutil==2.8.2

# 文本处理
jieba==0.42.1

# 开发和测试工具
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 可视化依赖
networkx==3.1
matplotlib==3.7.5

# 性能监控
psutil==5.9.6
memory-profiler==0.61.0

# 异步支持
aiohttp==3.8.6
aiofiles==23.2.1

# 工具库
click==8.1.7
tqdm==4.66.1
colorama==0.4.6

# 原有依赖保留
tensorboard==2.4.1
seqeval==1.2.2
opt-einsum==3.3.0
wandb==0.12.7
ujson==5.6.0
huggingface_hub>=0.14.1,<1.0
nltk==3.8
protobuf>=3.12.0,<=3.20.1
ipdb==0.13.11
pytorch-crf==0.7.2
openai==0.28.0
Jinja2==3.1.2
datasets==2.13.2
pyhocon==0.3.60
hydra-core>=1.1.0
zhconv>=1.4.0
paddlenlp>=2.4.0
rouge>=1.0.0
prettytable>=2.0.0
pyarrow>=5.0.0
xlrd2>=1.3.0
docx>=0.2.4
bmtrain>=0.2.0
deepspeed>=0.7.0
nlpaug>=1.1.0
tiktoken>=0.3.0
